requirements.txt
Babel==1.2
BeautifulSoup==3.2.1
Fabric==1.7.0
Flask==0.10.1
Flask-Admin==1.0.6
Flask-Babel==0.8
Flask-Cache==0.12
Flask-KVSession==0.4
Flask-Login==0.2.6
Flask-Mail==0.9.0
Flask-Misaka==0.2.0
Flask-Principal==0.4.0
Flask-Rauth==0.3.2
Flask-SQLAlchemy==1.0
Flask-Script==0.6.4
Flask-Security==1.6.7
Flask-WTF==0.8.4
Jinja2==2.7
Mako==0.8.1
MarkupSafe==0.18
PyGithub==1.17.0
Pygments==1.6
SQLAlchemy==0.8.2
Twisted==13.2.0
WTForms==1.0.4
Werkzeug==0.9.3
alembic==0.6.0
amqp==1.3.3
anyjson==0.3.3
autobahn==0.6.5
beautifulsoup4==4.3.2
billiard==3.3.0.13
blinker==1.3
celery==3.1.7
cql==1.4.0
cqlengine==0.9.2
distribute==0.7.3
docutils==0.11
fabtools==0.15.0
flower==0.5.2
httplib2==0.8
itsdangerous==0.22
kafka==0.8.1-3
kombu==3.0.8
misaka==1.0.2
msgpack-python==0.4.0
newrelic==1.13.1.31
oauth2==1.5.211
paramiko==1.11.0
passlib==1.6.1
poolbase==0.1.3
psycopg2==2.5.1
pycrypto==2.6
python-dateutil==2.1
python-memcached==1.53
pytz==2013b
rauth==0.6.1
raven==3.4.1
redis==2.7.6
requests==1.2.3
simplejson==3.3.1
simplekv==0.6
six==1.3.0
speaklater==1.3
thrift==0.9.1
tldextract==1.3.1
tornado==3.1
tweepy==2.1
uWSGI==1.9.14
wsgiref==0.1.2
zope.interface==4.0.5
